Refactor: move simple vote parsing to runtime
