+++ programs/vote/src/vote_state/mod.rs
@@ -41,73 +41,6 @@ pub const MAX_EPOCH_CREDITS_HISTORY: usize = 64;
 // Offset of VoteState::prior_voters, for determining initialization status without deserialization
 const DEFAULT_PRIOR_VOTERS_OFFSET: usize = 82;
 
-#[derive(Debug, PartialEq)]
-pub enum VoteTransaction {
-    Vote(Vote),
-    VoteStateUpdate(VoteStateUpdate),
-}
-
-impl VoteTransaction {
-    pub fn slots(&self) -> Vec<Slot> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.clone(),
-            VoteTransaction::VoteStateUpdate(vote_state_update) => vote_state_update
-                .lockouts
-                .iter()
-                .map(|lockout| lockout.slot)
-                .collect(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.is_empty(),
-            VoteTransaction::VoteStateUpdate(vote_state_update) => {
-                vote_state_update.lockouts.is_empty()
-            }
-        }
-    }
-
-    pub fn hash(&self) -> Hash {
-        match self {
-            VoteTransaction::Vote(vote) => vote.hash,
-            VoteTransaction::VoteStateUpdate(vote_state_update) => vote_state_update.hash,
-        }
-    }
-
-    pub fn timestamp(&self) -> Option<UnixTimestamp> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.timestamp,
-            VoteTransaction::VoteStateUpdate(vote_state_update) => vote_state_update.timestamp,
-        }
-    }
-
-    pub fn last_voted_slot(&self) -> Option<Slot> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.last().copied(),
-            VoteTransaction::VoteStateUpdate(vote_state_update) => {
-                Some(vote_state_update.lockouts.back()?.slot)
-            }
-        }
-    }
-
-    pub fn last_voted_slot_hash(&self) -> Option<(Slot, Hash)> {
-        Some((self.last_voted_slot()?, self.hash()))
-    }
-}
-
-impl From<Vote> for VoteTransaction {
-    fn from(vote: Vote) -> Self {
-        VoteTransaction::Vote(vote)
-    }
-}
-
-impl From<VoteStateUpdate> for VoteTransaction {
-    fn from(vote_state_update: VoteStateUpdate) -> Self {
-        VoteTransaction::VoteStateUpdate(vote_state_update)
-    }
-}
-
 #[frozen_abi(digest = "Ch2vVEwos2EjAVqSHCyJjnN2MNX1yrpapZTGhMSCjWUH")]
 #[derive(Serialize, Default, Deserialize, Debug, PartialEq, Eq, Clone, AbiExample)]
 pub struct Vote {
