+++ runtime/src/lib.rs
@@ -59,7 +59,9 @@ mod system_instruction_processor;
 pub mod transaction_batch;
 pub mod transaction_cost_metrics_sender;
 pub mod vote_account;
+pub mod vote_parser;
 pub mod vote_sender_types;
+pub mod vote_transaction;
 pub mod waitable_condvar;
 
 #[macro_use]
