+++ runtime/src/vote_sender_types.rs
@@ -1,6 +1,6 @@
 use {
+    crate::vote_parser::ParsedVote,
     crossbeam_channel::{Receiver, Sender},
-    solana_vote_program::vote_transaction::ParsedVote,
 };
 
 pub type ReplayVoteSender = Sender<ParsedVote>;
