+++ sdk/src/transaction/sanitized.rs
@@ -61,6 +61,7 @@ impl SanitizedTransaction {
         };
 
         let is_simple_vote_tx = is_simple_vote_tx.unwrap_or_else(|| {
+            // TODO: Move to `vote_parser` runtime module
             let mut ix_iter = message.program_instructions_iter();
             ix_iter.next().map(|(program_id, _ix)| program_id) == Some(&crate::vote::program::id())
         });
